Update configuration handling
The configuration for the DateFormat extension is now handled in the user configuration
instead of custom files. This is better on many levels. The first one is that there is
no need for specific rights to save the configuration. The second is that there is less
I/O access. The third one is that there is less code to maintain thus reducing the
possibility for bugs.